+++ frontend/src/hooks/index.ts
@@ -1,7 +1,8 @@
 export { default as useSidebar } from './useSidebar';
-export { default as useModalClose } from './useModalClose';
 export { default as useGroupAccessCode } from './useGroupAccessCode';
 export { default as useSearchParamAndQuery } from './useSearchParamAndQuery';
 export { default as useEyeButton } from './useEyeButton';
+export { default as usePasswordValidation } from './usePasswordValidation';
 export * from './review';
 export * from './reviewGroup';
+export * from './modal';
