+++ frontend/src/hooks/modal/index.ts
@@ -1,0 +1,2 @@
+export { default as useModalClose } from './useModalClose';
+export { default as useModals } from './useModals';
