+++ frontend/src/hooks/modal/useModalClose.ts
